api/requirements.txt
Django>=3.0.5
mysqlclient==1.4.6
djangorestframework==3.11.0
django-filter==2.2.0
Pillow==7.1.1
django-cors-headers==3.2.1
numpy
pandas
matplotlib
scikit-learn